requirements.txt
langchain>=0.2.0
langchain-community>=0.2
sentence-transformers>=2.6.1
faiss-cpu>=1.8.0
streamlit>=1.35.0
transformers>=4.42.0
accelerate>=0.31.0
protobuf<4.0.0      
pypdf
python-dotenv
llama-cpp-python==0.2.78